+++ src/exactly_lib/definitions/logic.py
@@ -1,3 +1,4 @@
+from exactly_lib.test_case import reserved_words
 from exactly_lib.util.logic_types import Quantifier
 
 NOT_OPERATOR_NAME = '!'
@@ -24,7 +25,7 @@ CONSTANT_MATCHER = 'constant'
 ALL_QUANTIFIER_ARGUMENT = 'every'
 EXISTS_QUANTIFIER_ARGUMENT = 'any'
 
-QUANTIFICATION_SEPARATOR_ARGUMENT = ':'
+QUANTIFICATION_SEPARATOR_ARGUMENT = reserved_words.COLON
 
 QUANTIFIER_ARGUMENTS = {
     Quantifier.ALL: ALL_QUANTIFIER_ARGUMENT,