+++ test/exactly_lib_test/impls/instructions/assert_/contents_of_file/test_resources/abstract_syntax.py
@@ -1,3 +1,4 @@
+from exactly_lib.test_case import reserved_words
 from exactly_lib_test.impls.types.string_matcher.test_resources.abstract_syntaxes import StringMatcherNegationAbsStx
 from exactly_lib_test.test_resources.source.abstract_syntax import AbstractSyntax
 from exactly_lib_test.test_resources.source.token_sequence import TokenSequence
@@ -29,6 +30,8 @@ class InstructionArgumentsAbsStx(AbstractSyntax):
     def tokenization(self) -> TokenSequence:
         return TokenSequence.concat([
             self._path.tokenization(),
+            TokenSequence.optional_new_line(),
+            TokenSequence.singleton(reserved_words.COLON),
             TokenSequence.optional_new_line(),
             self._matcher.tokenization(),
         ])