+++ test/exactly_lib_test/impls/instructions/assert_/existence_of_file/test_resources/arguments_building.py
@@ -2,7 +2,7 @@ from abc import ABC
 from typing import List, Optional
 
 from exactly_lib.definitions import logic
-from exactly_lib.impls.instructions.assert_ import existence_of_file
+from exactly_lib.test_case import reserved_words
 from exactly_lib.util.logic_types import ExpectationType
 from exactly_lib_test.impls.types.file_matcher.test_resources.argument_building import FileMatcherArg
 from exactly_lib_test.tcfs.test_resources.path_arguments import PathArgument
@@ -33,7 +33,7 @@ class CompleteInstructionArg(Argument):
         ret_val += self.path.elements
 
         if self.file_matcher is not None:
-            ret_val += [existence_of_file.PROPERTIES_SEPARATOR]
+            ret_val += [reserved_words.COLON]
             ret_val += self.file_matcher.elements
 
         return ret_val