+++ CHANGELOG.md
@@ -54,6 +54,7 @@ with exception of a "0." prefix until version 1 is released.
  - Exit codes - Values of all non-zero exit codes
  - Instruction `conf`/`actor` - Remove "-" prefix from actor names
  - Instruction `setup`/`stdin` - Contents value is `STRING-SOURCE`
+ - Instruction `assert`/`contents` - Syntax - add `:` between path and matcher (breaking)
  - Builtin symbol - string `LINE_SEP` -> Rename to `OS_LINE_SEP`
  - Syntax of `instruction description` - Delimiter is backtick (`) (breaking)  
  - Phase `conf` - Invalid HDS path gives `VALIDATION_ERROR`